[feat] add test app interfaces
