+++ app/test_app/test_app_1.py
@@ -1,0 +1,5 @@
+from app.test_app.test_app_interface import ITestApp
+
+
+class TestApp1(ITestApp):
+    pass
