+++ app/test_app/test_app_2.py
@@ -1,0 +1,5 @@
+from app.test_app.test_app_interface import ITestApp
+
+
+class TestApp2(ITestApp):
+    pass
