+++ app/test_app/test_app_interface.py
@@ -1,0 +1,9 @@
+from abc import ABC, abstractmethod
+
+from app.basic_logic import BasicLogic
+
+
+class ITestApp(ABC):
+    @abstractmethod
+    def run(self, basic_logic: BasicLogic):
+        pass
